fastmcp/requirements.txt
mcp>=1.9.0
uvicorn>=0.30.0
redis>=5.0.0
rq>=1.15.1
httpx>=0.27.0
filetype>=1.2.0
aiofiles>=23.2.1
boto3>=1.34.0
google-api-python-client>=2.130.0
google-auth>=2.34.0
